fix(spacings): declare types within component file
