+++ packages/components/spacings/spacings-inset/src/inset.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { FC } from 'react';
 import { css } from '@emotion/react';
 import { customProperties as vars } from '@commercetools-uikit/design-system';
 import { filterDataAttributes } from '@commercetools-uikit/utils';
@@ -27,7 +27,7 @@ type TProps = {
   children?: React.ReactNode;
 };
 
-const Inset: React.FC<TProps> = (props) => (
+const Inset: FC<TProps> = (props) => (
   <div
     css={css`
       padding: ${getPadding(props.scale)};
