--- packages/components/spacings/spacings-stack/src/types.ts
@@ -1,17 +1,0 @@
-export type TAlignItem =
-  | 'stretch'
-  | 'flex-start'
-  | 'flex-end'
-  | 'center'
-  | 'baseline'
-  // Deprecated
-  | 'flexStart'
-  | 'flexEnd';
-
-export type TScale = 'xs' | 's' | 'm' | 'l' | 'xl';
-
-export type TProps = {
-  scale?: TScale;
-  alignItems?: TAlignItem;
-  children?: React.ReactNode;
-};
